Change File Structure
